Start work on adding a CAN Frame buffer logic.
The current plan is to send one byte at a time to the serial port and
between each byte check the CAN controller for a message. If there is
a message queue it in the buffer and go back to the serial port.

Eventually it needs to be written in a way that fully utilizes the
interrupts that are available on the AVR.